# PERF: Cythonize groupby transforms #4095