… pandas/tests/test_groupby.py