closes - #4095 - adding Cythonized versions of `shift`, `cumprod` and `cumsum`, along with a helper ``GroupBy._cython_transform`` that other cythonized transform-ops could be added to.

I'm using fused types here instead of adding to `generated.pyx` - as far as I can tell that is working well, looking back through the old issues (#255) it looks like that wasn't done originally because of older versions of Cython?

### 0.16.2

    In [17]: df = pd.DataFrame({'i':range(1000) * 1000, 
                                'f':np.linspace(0, 2000, 1000000),    
                                'f2':np.linspace(1, 2, 1000000)})

    In [18]: %timeit df.groupby('i').cumsum()
    1 loops, best of 3: 614 ms per loop

    In [19]: %timeit df.groupby('i').shift()
    1 loops, best of 3: 499 ms per loop


### PR

    In [2]: %timeit df.groupby('i').cumsum()
    10 loops, best of 3: 58.6 ms per loop

    In [3]: %timeit df.groupby('i').shift()
    10 loops, best of 3: 73.4 ms per loop