@chris-b1 2015-08-27
Travis build is failing with the message below - which seems like it might be an out of memory error?

     gcc: internal compiler error: Killed (program cc1)
     Please submit a full bug report, with preprocessed source if appropriate.