@chris-b1 2015-09-04
@jreback - any guesses on why the Travis build is failing?  It builds fine locally on windows, also worked in a linux vm.