@jreback 2015-09-04
I think this doesn't build under cython 0.19.1 (our min version).