@jreback 2015-09-04
looks pretty odd: https://travis-ci.org/pydata/pandas/jobs/77861783, ``gcc kill``, no idea :<