@kawochen 2015-09-07
This could indicate a gcc memory problem.  Refactoring into smaller files should make it easier to compile.