@jreback 2015-10-11
need some asv benches for this