@chris-b1 2015-10-11
@jreback - updated for your comments, moved the cython into generated.pyx, added some benches, and refactored `BaseGrouper` to put the cythonized `transform` methods alongside `aggregation`.

```
  [b2a547e8] [d94947e]
-    10.37s     8.38ms      0.00  groupby.groupby_transform_cythonized.time_groupby_transform_shift
-    13.68s     6.76ms      0.00  groupby.groupby_transform_cythonized.time_groupby_transform_cumsum
-    13.59s     6.70ms      0.00  groupby.groupby_transform_cythonized.time_groupby_transform_cumprod
```