@jreback 2015-10-14
this should always be a float calc