@chris-b1 2015-10-16
@jreback, if you want to take another look, I think I've got your comments addressed.  

The reference tests (e.g. `gb.cumsum() ==  gb.apply(lambda x:x.cumsum()`) have gotten pretty complex - I probably need to add some simpler ones with hard-coded output too.