@jreback 2015-11-13
@chris-b1 ok, looks reasonable. pls rebase / squash. ping when green.