@chris-b1 2015-11-15
@jreback - this is updated (also added whatsnew) - let me know if it needs anything else.