@jreback 2015-11-16
merged via https://github.com/pydata/pandas/commit/b07dd0cbd6d18c55aaa0043d85f42a483eab7dbb

thanks!